Moved all parts of code from main.js into their own modules which can be called.
