+++ modules/voicesystem.js
@@ -1,0 +1,4 @@
+module.exports = {
+   main: function(message) {
+
+   }
