+++ versioninfo.json
@@ -1,0 +1,6 @@
+{
+  "mainVersion": "1.0.3",
+  "commandsframeworkVersion": "1.0.1",
+  "loggingframeworkVersion": "1.0.2",
+  "statstrackVersion": "1.0.3"
+}
